Replaced SQL call with API call
Removed a query to the V$SESSION table as it was not necessary.  The package DBMS_APPLICATION_INFO has an api call which removes the need to query the session information dba view directly.  Propogated the required changes to the CAPTURE_SESSION procedure and the LOG procedure.  Also removed references to SESSION_ID where it was no longer needed.